ml-pipeline/data-loader/requirements.txt
mysql-connector-python==9.3.0
boto3==1.38.38